Extract domain settings into own module
Move the custom-domain and host-subdomain handlers out of settings.ts
into a dedicated settings-domains.ts, trimming the bunny-cdn import to
the one name the page state still needs and dropping the now single-use
DOMAIN_PATTERN import.
